refactor: isolate runtime ecstore facade imports
