+++ crates/obs/src/metrics/mod.rs
@@ -14,26 +14,17 @@
 
 pub mod collectors;
 pub mod config;
+mod ecstore_compat;
 pub mod report;
 pub mod scheduler;
 pub mod schema;
 pub mod stats_collector;
 
-pub(crate) use rustfs_ecstore::api::bucket::lifecycle::bucket_lifecycle_ops::{
-    GLOBAL_ExpiryState as OBS_GLOBAL_EXPIRY_STATE, GLOBAL_TransitionState as OBS_GLOBAL_TRANSITION_STATE,
+pub(crate) use ecstore_compat::{
+    OBS_GLOBAL_EXPIRY_STATE, OBS_GLOBAL_REPLICATION_STATS, OBS_GLOBAL_TRANSITION_STATE, ObsEcstoreResult, ObsStore,
+    obs_get_global_bucket_monitor, obs_get_quota_config, obs_get_total_usable_capacity, obs_get_total_usable_capacity_free,
+    obs_load_data_usage_from_backend, obs_resolve_object_store_handle,
 };
-pub(crate) use rustfs_ecstore::api::bucket::metadata_sys::get_quota_config as obs_get_quota_config;
-pub(crate) use rustfs_ecstore::api::bucket::replication::GLOBAL_REPLICATION_STATS as OBS_GLOBAL_REPLICATION_STATS;
-pub(crate) use rustfs_ecstore::api::capacity::{
-    get_total_usable_capacity as obs_get_total_usable_capacity,
-    get_total_usable_capacity_free as obs_get_total_usable_capacity_free,
-};
-pub(crate) use rustfs_ecstore::api::data_usage::load_data_usage_from_backend as obs_load_data_usage_from_backend;
-pub(crate) use rustfs_ecstore::api::error::Result as ObsEcstoreResult;
-pub(crate) use rustfs_ecstore::api::global::{
-    get_global_bucket_monitor as obs_get_global_bucket_monitor, resolve_object_store_handle as obs_resolve_object_store_handle,
-};
-pub(crate) use rustfs_ecstore::api::storage::ECStore as ObsStore;
 
 pub use collectors::*;
 pub use config::*;
