+++ crates/s3select-api/src/lib.rs
@@ -13,20 +13,10 @@
 // limitations under the License.
 
 use datafusion::{common::DataFusionError, sql::sqlparser::parser::ParserError};
-use rustfs_ecstore::api::error::{
-    is_err_bucket_not_found as select_is_err_bucket_not_found_from_backend,
-    is_err_object_not_found as select_is_err_object_not_found_from_backend,
-    is_err_version_not_found as select_is_err_version_not_found_from_backend,
-};
-use rustfs_ecstore::api::global::resolve_object_store_handle as resolve_select_object_store_handle_from_backend;
 use snafu::{Backtrace, Location, Snafu};
 use std::fmt::Display;
-use std::sync::Arc;
 
-pub(crate) use rustfs_ecstore::api::error::StorageError as SelectStorageError;
-pub(crate) use rustfs_ecstore::api::set_disk::DEFAULT_READ_BUFFER_SIZE as SELECT_DEFAULT_READ_BUFFER_SIZE;
-pub(crate) use rustfs_ecstore::api::storage::ECStore as SelectStore;
-
+mod ecstore_compat;
 pub mod object_store;
 pub mod query;
 pub mod server;
@@ -36,25 +26,11 @@ mod test;
 
 pub type QueryResult<T> = Result<T, QueryError>;
 
-pub(crate) type SelectGetObjectReader = <SelectStore as rustfs_storage_api::ObjectIO>::GetObjectReader;
-pub(crate) type SelectObjectInfo = <SelectStore as rustfs_storage_api::ObjectOperations>::ObjectInfo;
-pub(crate) type SelectObjectOptions = <SelectStore as rustfs_storage_api::ObjectOperations>::ObjectOptions;
-
-pub(crate) fn resolve_select_object_store_handle() -> Option<Arc<SelectStore>> {
-    resolve_select_object_store_handle_from_backend()
-}
-
-pub(crate) fn select_is_err_bucket_not_found(err: &SelectStorageError) -> bool {
-    select_is_err_bucket_not_found_from_backend(err)
-}
-
-pub(crate) fn select_is_err_object_not_found(err: &SelectStorageError) -> bool {
-    select_is_err_object_not_found_from_backend(err)
-}
-
-pub(crate) fn select_is_err_version_not_found(err: &SelectStorageError) -> bool {
-    select_is_err_version_not_found_from_backend(err)
-}
+pub(crate) use ecstore_compat::{
+    SELECT_DEFAULT_READ_BUFFER_SIZE, SelectGetObjectReader, SelectObjectInfo, SelectObjectOptions, SelectStorageError,
+    SelectStore, resolve_select_object_store_handle, select_is_err_bucket_not_found, select_is_err_object_not_found,
+    select_is_err_version_not_found,
+};
 
 #[derive(Debug, Snafu)]
 #[snafu(visibility(pub))]
