+++ docs/architecture/migration-progress.md
@@ -5,18 +5,18 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 ## Current Context
 
 - Issue: [`rustfs/backlog#660`](https://github.com/rustfs/backlog/issues/660)
-- Branch: `overtrue/arch-admin-ecstore-owner-symbols`
-- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100/API-101/API-102/API-103/API-104/API-105/API-106/API-107/API-108/API-109/API-110/API-111/API-112/API-113/API-114/API-115/API-116/API-117/API-118/API-119/API-120/API-121/API-122/API-123/API-124/API-125/API-126/API-127/API-128/API-129/API-130/API-131/API-132/API-133/API-134/API-135/API-136/API-137/API-138/API-139/API-140/API-141/API-142/API-143/API-144/API-145`.
-- Based on: API-145 stacked slice.
+- Branch: `overtrue/arch-runtime-crate-ecstore-boundaries`
+- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100/API-101/API-102/API-103/API-104/API-105/API-106/API-107/API-108/API-109/API-110/API-111/API-112/API-113/API-114/API-115/API-116/API-117/API-118/API-119/API-120/API-121/API-122/API-123/API-124/API-125/API-126/API-127/API-128/API-129/API-130/API-131/API-132/API-133/API-134/API-135/API-136/API-137/API-138/API-139/API-140/API-141/API-142/API-143/API-144/API-145/API-146`.
+- Based on: API-146 stacked slice.
 - PR type for this branch: `pure-move`
 - Runtime behavior changes: none.
-- Rust code changes: replace RustFS admin facade ECStore source imports with
-  storage-owner re-exports.
+- Rust code changes: move external runtime crate ECStore source imports into
+  crate-local `ecstore_compat` boundaries.
 - CI/script changes: lock completed owner and test/fuzz boundaries against
   bare/glob imports, scattered raw ECStore facade subpaths, and startup
   runtime/root-server/table/S3/app shared/app bucket/app ECStore/admin facade
-  regressions.
-- Docs changes: record the API-136/API-137/API-138/API-139/API-140/API-141/API-142/API-143/API-144/API-145/API-146 owner facade cleanup.
+  regressions, plus external runtime ECStore compatibility bypasses.
+- Docs changes: record the API-136/API-137/API-138/API-139/API-140/API-141/API-142/API-143/API-144/API-145/API-146/API-147 owner facade cleanup.
 
 ## Phase 0 Tasks
 
@@ -4117,6 +4117,23 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
     syntax check, formatting, diff hygiene, Rust risk scan, branch freshness
     check, pre-commit, and three-expert review.
 
+- [x] `API-147` Route external runtime crate ECStore source imports through local compatibility boundaries.
+  - Do: move direct ECStore facade source imports in notify, observability
+    metrics, S3 Select, Swift, IAM, heal, and scanner runtime entry modules
+    into crate-local `ecstore_compat` modules while preserving existing
+    wrappers and aliases at each crate boundary.
+  - Acceptance: target runtime crate source directories contain no direct
+    `rustfs_ecstore::api::` source paths outside `ecstore_compat.rs`; migration
+    guards reject restoring those bypasses.
+  - Must preserve: notify config persistence and object-store resolution,
+    observability storage/ILM/replication metrics, S3 Select storage error
+    mapping and object reads, Swift metadata/object-store access, IAM config
+    and notification behavior, heal disk wrappers, and scanner lifecycle/disk
+    runtime wrappers.
+  - Verification: focused external crate compile, migration guard, shell syntax
+    check, formatting, diff hygiene, Rust risk scan, branch freshness check,
+    pre-commit, and three-expert review.
+
 ## Next PRs
 
 1. `pure-move`: continue pruning remaining facade compatibility and owner boundaries.
@@ -4125,9 +4142,9 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 
 | Expert | Status | Notes |
 |---|---|---|
-| Quality/architecture | pass | API-146 clears direct ECStore source imports from the admin facade and guards the completed admin boundary. |
-| Migration preservation | pass | Admin facade shape stays unchanged; bucket controls, config, data usage, global metadata, metrics, notification, rebalance, RPC, storage, and tier admin paths still delegate to the same ECStore backend modules through storage. |
-| Testing/verification | pass | Focused RustFS test-target compile, shell syntax, migration guard, formatting, diff hygiene, Rust risk scan, and pre-commit passed for API-146. |
+| Quality/architecture | pass | API-147 moves selected external runtime crate ECStore source imports behind crate-local compatibility files and guards those boundaries. |
+| Migration preservation | pass | Notify, OBS metrics, S3 Select, Swift, IAM, heal, and scanner root APIs keep the same aliases/wrappers while only the ECStore source location changes. |
+| Testing/verification | pass | Focused external crate compile, formatting, migration guard, diff hygiene, Rust risk scan, and pre-commit passed for API-147. |
 
 ## Verification Notes
 
@@ -4215,6 +4232,21 @@ Passed before push:
   - Rust risk scan: no new production unwrap/expect, casts, panic/todo/unsafe,
     or error-type risks added; only existing `DiskResult<Vec<String>>`
     textual matches were reported by the broad error-type scan.
+
+- Issue #660 API-147 current slice:
+  - `cargo check -p rustfs-notify -p rustfs-obs -p rustfs-s3select-api -p rustfs-protocols -p rustfs-iam -p rustfs-heal -p rustfs-scanner`:
+    passed.
+  - `cargo fmt --all`: passed.
+  - `cargo fmt --all --check`: passed.
+  - `git diff --check`: passed.
+  - `bash -n scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_architecture_migration_rules.sh`: passed.
+  - `make pre-commit`: passed.
+  - Runtime crate ECStore source bypass scan: passed; target runtime crate
+    source paths now reference `rustfs_ecstore::api::` only inside
+    `ecstore_compat.rs`.
+  - Rust risk scan: no new production unwrap/expect, casts, panic/todo/unsafe,
+    or error-type risks added; changes only move import/source boundaries.
 
 - Issue #660 API-139 current slice:
   - `cargo check --tests -p rustfs`: passed.
